+++ .pylintrc
@@ -100,3 +100,8 @@ msg-template='{path}:{line}: [{msg_id}({symbol}), {obj}] {msg}'
 # R0801: duplicate code across files
 #        (We have a lot of intentional duplication, like the tutorials)
 disable=C0103,C0111,W0613,E1101,E0611,R0913,R0914,W0223,E0203,W0201,C1801,E1129,C0325,R1705,W0622,W0703,W0101,W0122,R0912,R0903,W0221,W0212,R1703,R0201,W0603,R0915,R0902,C0200,W1201,C0302,W0511,R0911,E0401,R0205,R1714,R0801,E1130
+
+[IMPORTS]
+# Explicitly specifying this helps pylint to behave more consistently across multiple platforms.
+# If not specified, the method used to install each library seems to change whether it is considered third party, etc.
+known-third-party=joblib,keras,numpy,PIL,six,tensorflow
